extend IFOPT script to add missing stops
This changes the script that adds IFOPT numbers to the Nominatim
to also add artifical nodes for all stops that cannot be found at
all in the OSM data in Nominatim.

The script tries to be clever about existing artificial stops and
only add the when there is a difference between the data in the
database and the CSV. So it should be possible to update the
reference CSV with minimal impact on the database.

The script uses node IDs that are currently not used in OSM. It
assumes that it is the only user of that ID space. It won't play
well with other users trying to inject artifical nodes.